move interface to its own file
