+++ control.lua
@@ -429,12 +429,6 @@ local function remove_spider(registration_number)
   global.spidertrons[registration_number] = nil
 end
 
----@param name string
-local function ignore_spider(name)
-  global.ignored_spidertrons = global.ignored_spidertrons or {}
-  global.ignored_spidertrons[name] = true
-end
-
 local function initialize_globals()
   global.spidertrons = {}
   for _, surface in pairs(game.surfaces) do
@@ -462,11 +456,7 @@ local function on_entity_destroyed(event)
   remove_spider(event.registration_number)
 end
 
-local interface_functions = {
-  ignore_spider = ignore_spider,
-}
-remote.add_interface("wandering-spiders", interface_functions)
--- usage: remote.call("wandering-spiders", "ignore_spider", "name of spider to ignore")
+require("interface")
 
 script.on_init(initialize_globals)
 script.on_configuration_changed(initialize_globals)
